pyproject.toml
[project]
name = "pchome-tracker"
version = "0.1.0"
description = "PChome 24h tracking list price follower"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
    "ty>=0.0.1a6",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
